serving/seldon_server/requirements.txt
seldon-core==1.18.2
uvicorn===0.31.0
torch==2.4.0
torchvision==0.19.0
transformers==4.44.2
supervision==0.22.0
huggingface==0.0.1
accelerate==0.33.0
torchmetrics==1.4.1
albumentations==1.4.14
pillow==10.4.0
datasets==2.21.0
PyYAML==6.0.2
wandb==0.17.7
pytest==8.3.3
pycocotools==2.0.8
python-dotenv==1.0.1
boto3==1.34.158